+++ CHANGELOG.md
@@ -7,8 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 
 ### Fixed 
-- Fixed guide description in Recently Viewer [#322](https://github.com/rokwire/illinois-app/issues/322).
+- Fixed guide description in Recently Viewed [#322](https://github.com/rokwire/illinois-app/issues/322).
 - Do not show favorite button in Student Guide items if privacy level does not support this.
+- Strip HTML tags from guide list title & description when show them in Saved or Recently Viewed.
 - Events and virtual events improvements [#321](https://github.com/rokwire/illinois-app/issues/321).
 
 ## [2.4.27] - 2021-06-21